speculation-rules/prefetch/duplicate-urls.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js?feature=bidi"></script>
<script src="/resources/testdriver-extra.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>

<script>
  setup(() => assertSpeculationRulesIsSupported());

  promise_test(async t => {
    await test_driver.bidi.speculation.prefetch_status_updated.subscribe();
    let statusPromise = new Promise(resolve => {
      test_driver.bidi.speculation.prefetch_status_updated.on(event => {
        if (event && event.status !== undefined) resolve(event.status);
      });
    });
    let urls = Array(5).fill(getPrefetchUrlList(1)[0]);
    insertSpeculationRules({ prefetch: [{ source: 'list', urls: urls }] });
    await statusPromise;
    await new Promise(resolve => t.step_timeout(resolve, 2000));

    let prefetched_count = await isUrlPrefetched(urls[0]);
    assert_equals(prefetched_count, 1, "url should be prefetched just once.");
  }, "browser should remove duplicate urls from prefetch buffer.");
</script>
